Add ListCRUDProvider
