+++ .changeset/serious-news-peel.md
@@ -1,0 +1,5 @@
+---
+'@keystonejs/keystone': patch
+---
+
+Added a ListCRUDProvider to generate the CRUD graphQL API for lists.
