+++ packages/keystone/lib/Keystone/index.js
@@ -1,4 +1,3 @@
-const { GraphQLJSON } = require('graphql-type-json');
 const fs = require('fs');
 const gql = require('graphql-tag');
 const flattenDeep = require('lodash.flattendeep');
@@ -36,7 +35,12 @@ const {
 } = require('./relationship-utils');
 const List = require('../List');
 const { DEFAULT_DIST_DIR } = require('../../constants');
-const { CustomProvider, ListAuthProvider, VersionProvider } = require('../providers');
+const {
+  CustomProvider,
+  ListAuthProvider,
+  ListCRUDProvider,
+  VersionProvider,
+} = require('../providers');
 
 const debugGraphQLSchemas = () => !!process.env.DEBUG_GRAPHQL_SCHEMAS;
 
@@ -76,12 +80,17 @@ module.exports = class Keystone {
     this._schemaNames = schemaNames;
     this.appVersion = appVersion;
 
+    this._listCRUDProvider = new ListCRUDProvider();
     this._customProvider = new CustomProvider({
       schemaNames,
       defaultAccess: this.defaultAccess,
       buildQueryHelper: this._buildQueryHelper,
     });
-    this._providers = [this._customProvider, new VersionProvider({ appVersion, schemaNames })];
+    this._providers = [
+      this._listCRUDProvider,
+      this._customProvider,
+      new VersionProvider({ appVersion, schemaNames }),
+    ];
 
     if (adapters) {
       this.adapters = adapters;
@@ -321,6 +330,7 @@ module.exports = class Keystone {
     });
     this.lists[key] = list;
     this.listsArray.push(list);
+    this._listCRUDProvider.lists.push(list);
     list.initFields();
     return list;
   }
@@ -413,100 +423,17 @@ module.exports = class Keystone {
   }
 
   getTypeDefs({ schemaName }) {
-    // Aux lists are only there for typing and internal operations, they should
-    // not have any GraphQL operations performed on them
-    const firstClassLists = this.listsArray.filter(list => !list.isAuxList);
-
-    const mutations = unique(
-      flatten([
-        ...firstClassLists.map(list => list.getGqlMutations({ schemaName })),
-        ...this._providers.map(p => p.getMutations({ schemaName })),
-      ])
-    );
-
+    const queries = unique(flatten(this._providers.map(p => p.getQueries({ schemaName }))));
+    const mutations = unique(flatten(this._providers.map(p => p.getMutations({ schemaName }))));
     // Fields can be represented multiple times within and between lists.
     // If a field defines a `getGqlAuxTypes()` method, it will be
     // duplicated.
     // graphql-tools will blow up (rightly so) on duplicated types.
     // Deduping here avoids that problem.
     return [
-      ...unique(flatten(this.listsArray.map(list => list.getGqlTypes({ schemaName })))),
       ...unique(flatten(this._providers.map(p => p.getTypes({ schemaName })))),
-      `"""NOTE: Can be JSON, or a Boolean/Int/String
-          Why not a union? GraphQL doesn't support a union including a scalar
-          (https://github.com/facebook/graphql/issues/215)"""
-       scalar JSON`,
-      `type _ListAccess {
-          """Access Control settings for the currently logged in (or anonymous)
-             user when performing 'create' operations.
-             NOTE: 'create' can only return a Boolean.
-             It is not possible to specify a declarative Where clause for this
-             operation"""
-          create: Boolean
-
-          """Access Control settings for the currently logged in (or anonymous)
-             user when performing 'read' operations."""
-          read: JSON
-
-          """Access Control settings for the currently logged in (or anonymous)
-             user when performing 'update' operations."""
-          update: JSON
-
-          """Access Control settings for the currently logged in (or anonymous)
-             user when performing 'delete' operations."""
-          delete: JSON
-
-          """Access Control settings for the currently logged in (or anonymous)
-             user when performing 'auth' operations."""
-          auth: JSON
-        }`,
-      `type _ListSchemaRelatedFields {
-        """The typename as used in GraphQL queries"""
-        type: String
-
-        """A list of GraphQL field names"""
-        fields: [String]
-      }`,
-      `type _ListSchema {
-        """The typename as used in GraphQL queries"""
-        type: String
-
-        """Top level GraphQL query names which either return this type, or
-           provide aggregate information about this type"""
-        queries: [String]
-
-        """Information about fields on other types which return this type, or
-           provide aggregate information about this type"""
-        relatedFields: [_ListSchemaRelatedFields]
-      }`,
-      `type _ListMeta {
-        """The Keystone List name"""
-        name: String
-
-        """Access control configuration for the currently authenticated
-           request"""
-        access: _ListAccess
-
-        """Information on the generated GraphQL schema"""
-        schema: _ListSchema
-       }`,
-      `type _QueryMeta {
-          count: Int
-       }`,
-      `type Query {
-          ${unique(
-            flatten([
-              ...firstClassLists.map(list => list.getGqlQueries({ schemaName })),
-              ...this._providers.map(p => p.getQueries({ schemaName })),
-            ])
-          ).join('\n')}
-          """ Retrieve the meta-data for all lists. """
-          _ksListsMeta: [_ListMeta]
-       }`,
-      mutations.length > 0 &&
-        `type Mutation {
-          ${mutations.join('\n')}
-       }`,
+      queries.length > 0 && `type Query { ${queries.join('\n')} }`,
+      mutations.length > 0 && `type Mutation { ${mutations.join('\n')} }`,
     ]
       .filter(s => s)
       .map(s => print(gql(s)));
@@ -524,54 +451,6 @@ module.exports = class Keystone {
     if (debugGraphQLSchemas()) {
       typeDefs.forEach(i => console.log(i));
     }
-
-    const queryMetaResolver = {
-      // meta is passed in from the list's resolver (eg; '_allUsersMeta')
-      count: meta => meta.getCount(),
-    };
-
-    const listMetaResolver = {
-      // meta is passed in from the list's resolver (eg; '_allUsersMeta')
-      access: meta => meta.getAccess(),
-      // schema is
-      schema: meta => meta.getSchema(),
-    };
-
-    const listAccessResolver = {
-      // access is passed in from the listMetaResolver
-      create: access => access.getCreate(),
-      read: access => access.getRead(),
-      update: access => access.getUpdate(),
-      delete: access => access.getDelete(),
-      auth: access => access.getAuth(),
-    };
-
-    // Aux lists are only there for typing and internal operations, they should
-    // not have any GraphQL operations performed on them
-    const firstClassLists = this.listsArray.filter(list => !list.isAuxList);
-
-    // NOTE: some fields are passed through unchanged from the list, and so are
-    // not specified here.
-    const listSchemaResolver = {
-      // A function so we can lazily evaluate this potentially expensive
-      // operation
-      // (Could we memoize this in the future?)
-      // NOTE: We purposely include the list we're looking for as it may have a
-      // self-referential field (eg: User { friends: [User] })
-      relatedFields: ({ key }) =>
-        firstClassLists
-          .map(list => ({
-            type: list.gqlNames.outputTypeName,
-            fields: flatten(
-              list
-                .getFieldsRelatedTo(key)
-                .filter(field => field.access[schemaName].read)
-                .map(field => Object.keys(field.gqlOutputFieldResolvers({ schemaName })))
-            ),
-          }))
-          .filter(({ fields }) => fields.length),
-    };
-
     // Like the `typeDefs`, we want to dedupe the resolvers. We rely on the
     // semantics of the JS spread operator here (duplicate keys are overridden
     // - first one wins)
@@ -582,36 +461,9 @@ module.exports = class Keystone {
       {
         // Order of spreading is important here - we don't want user-defined types
         // to accidentally override important things like `Query`.
-        ...objMerge(this.listsArray.map(list => list.gqlAuxFieldResolvers({ schemaName }))),
-        ...objMerge(this.listsArray.map(list => list.gqlFieldResolvers({ schemaName }))),
-
-        JSON: GraphQLJSON,
-
-        _QueryMeta: queryMetaResolver,
-        _ListMeta: listMetaResolver,
-        _ListAccess: listAccessResolver,
-        _ListSchema: listSchemaResolver,
-
         ...objMerge(this._providers.map(p => p.getTypeResolvers({ schemaName }))),
-
-        Query: {
-          // Order is also important here, any TypeQuery's defined by types
-          // shouldn't be able to override list-level queries
-          ...objMerge(firstClassLists.map(list => list.gqlAuxQueryResolvers())),
-          ...objMerge(firstClassLists.map(list => list.gqlQueryResolvers({ schemaName }))),
-          // And the Keystone meta queries must always be available
-          _ksListsMeta: (_, args, context) =>
-            this.listsArray
-              .filter(list => list.access[schemaName].read)
-              .map(list => list.listMeta(context)),
-          ...objMerge(this._providers.map(p => p.getQueryResolvers({ schemaName }))),
-        },
-
-        Mutation: {
-          ...objMerge(firstClassLists.map(list => list.gqlAuxMutationResolvers())),
-          ...objMerge(firstClassLists.map(list => list.gqlMutationResolvers({ schemaName }))),
-          ...objMerge(this._providers.map(p => p.getMutationResolvers({ schemaName }))),
-        },
+        Query: objMerge(this._providers.map(p => p.getQueryResolvers({ schemaName }))),
+        Mutation: objMerge(this._providers.map(p => p.getMutationResolvers({ schemaName }))),
       },
       o => Object.entries(o).length > 0
     );
