+++ packages/keystone/lib/providers/index.js
@@ -1,5 +1,6 @@
 const { CustomProvider } = require('./custom');
 const { ListAuthProvider } = require('./listAuth');
+const { ListCRUDProvider } = require('./listCRUD');
 const { VersionProvider } = require('./version');
 
 // The GraphQL Provider Framework expects to see classes with the following API:
@@ -28,4 +29,4 @@ const { VersionProvider } = require('./version');
 //   }
 // }
 
-module.exports = { CustomProvider, ListAuthProvider, VersionProvider };
+module.exports = { CustomProvider, ListAuthProvider, ListCRUDProvider, VersionProvider };
